+++ src/stories/Introduction.stories.mdx
@@ -108,7 +108,5 @@ import { version } from '../../package.json';
   `}
 </style>
 
-<>
-	<h1>Welcome to Lunatic storybook version: {version} </h1>
-	<img src={Lunatic} alt="Lunatic logo" />
-</>
+<h1>Welcome to Lunatic storybook version: {version} </h1>
+<img src={Lunatic} alt="Lunatic logo" />
